+++ mjx/config.yaml
@@ -12,7 +12,7 @@ num_updates_per_batch: 8
 discounting: 0.97
 learning_rate: 0.0003
 entropy_cost: 0.001
-num_envs: 1
+num_envs: 4096
 # num_envs: 1
 batch_size: 512
 seed: 0
