--- weights/blank.pkl
@@ -1,1 +1,0 @@
-Model should be saved in the weights folder with name "model.pkl"
